refactor: create email and password validation services for login
create email and password validation services for login

MODULE-1-PART-3-gallery
