+++ login/login.js
@@ -1,4 +1,4 @@
-import { validateUserInput } from "./validate-handler.js";
+import { validateUserInput } from "./validate.handler.js";
 import { submitUserData } from "../utilities/dataSubmitting/dataSubmitting.js";
 
 const loginForm = document.getElementById('loginForm');
